completed flask routes

+++ wallet.py
@@ -3,7 +3,7 @@ from app import db
 class Wallet(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     userID = db.Column(db.Integer,db.ForeignKey('user.id'))
-    amount = db.Column(db.Integer,default=0)
+    amount = db.Column(db.Float,default=0)
     currencyType = db.Column(db.String()) #change to ID from foreign key if there is time
     def __init__(self,userID,currencyType,amount):
         self.userID = userID
@@ -14,6 +14,7 @@ class Wallet(db.Model):
     def __repr__(self):
         return '<id {}>'.format(self.id)
 
+   
     def serialize(self):
         return {
             'id': self.id,
